Add constructors, set+get to model classes
Also enable add/remove to Lists or Sets (where they appear).
It is the Project's responsibility to establish the bidirectional links
it has with the images and the tags.
Thus, using the .addImage, .addTag etc methods of a Project instance
suffices for establishing the link.

Also added annotations for JSON de/serialization