Add a WebSecurityConfig
This is basically to enable the DaoAuthenticationProvider which in
turn uses a PasswordEncoder to simplify password handling
(technically, we just need to encode when the password is set;
password decoding/matching is done via the Provider)
We use BCrypt as the cryptographic algorithm.

Of course, we will later use the WebSecurityConfig.configure(http)
method to require users to be authenticated on some endpoints.