Add a UserController
It has the POST '/api/signup' route where registering will take
place.
It also has the GET '/api/user/{id}' route where may access public
data about a single user.

It uses Spring HATEOAS Link-builder to generate the Location
header in the 201 status response when successfully creating a
new User (that is, it automatically generates a URI at which the
new User can be accessed).
That same User's PublicData is returned in the response
body.

It also handles the validation errors on the constraints on the
SignUpRequest object and returns a 400 status with a
{incorrectField}: {errorMessage} dictionary as body.

It also handles the UserAlreadyExistException in a similar
fashion, but returns a 403 status.