Migrate old JS files to TS
.js -> .jsx -> .tsx if it imports/needs React
.js -> .ts otherwise

Used errors thrown by 'tsc' to inform our fixes
There is no immediate clean solution but to declare React's props
as being of type 'any'.
If a prop really needs a field, it can be declared as such in the
future...

Types are not as much necessary for the machine, as they are to
prevent human error.