Add the InputField component
It is colored red (.is-danger) if the input field is invalid or
green (.is-success) if it is valid and it has been touched.
It is plain otherwise.

It has an inner portion for the ValidationIcon and an outer
portion for the ErrorMessageDisplay

All of the above conditional rendering is done through props