Add the Formik form in SignUpPage component
The SignUpSchema is also here.
Note that the keys of schema object have the same name as the fields
in SignUpRequest in our server-side model!

Formik basically takes a map of initial values and uses its keys to
index the fields (and their corresponding errors+touched status)
and also uses its values to validate the Yup SignUpSchema
On submit, the map with the updated values is processed
(for now they are just logged to console)

It is bulky and should be refactored, but type constraints will
not allow something as simple as 'Object.keys(errors).map()'

Also, the password fields require the password type for the input
and should probably be differentiated somehow.