Update SignUpPage to attempt to submit form
We made it so that it is a React function component (and not a class)
since we will be using hooks.

We use a Jotai boolean atom to track if the SignUp should succeed
(which we asume to be true by default). If we get a 403 from the
server, then the username is not unique and we should not leave
the page (we should not discard the state with FormikValues), as
we would in the case of a 201 (CREATED) response

Function attemptToSubmit(values) is not added yet...