Add rest of SignUpPage logic
We add another ErrorMessageDisplay only for the 'username' InputField.
This one hasError only if the Jotai atom discussed before (here
labeled as 'isUniqueUser') is false: it is an error distinct from the
Yup SignUpSchema.

The large attemptToSubmit(values) function was further broken down
into a handleSignupResponse(response, dataToPersist). The latter is a
switch based on the response.status

dataToPersist is data from the form itself (username + Base64 encoded
password) that needs to be saved client-side (for BasicAuth) if the
response was successful.

We also add a return to mainpage button and some CSS classes on the
buttons