Fix display of cards on MainPage
Now, we maintain the responsiveness of flex, but limit its growth
up to max. 3 cards per row via 'flex-basis' for the children (the
cards themselves)

Also, images now take up the whole parent div space within a card
This is done via 'min-width' and 'translate' (so that it is centered
when it is cropped by 'display: block'). Note that the 5-by-3 ratio
is maintained thanks to Bulma.

As a side effect, we got rid of some of the prop drilling via
'props.width'