Refactor+Add components on ProjectPage
There is still work to be done...
The 55vw used in the child components (such as BigImage) has been
exported as a constant to have it synced here.

The whole page uses this unit to display a centered column.

We will also use a SocialMediaContainer to hold the Share button and
the Tags for the project.