Add the LoginForm nested in the LoginModal
Again, this is copy+paste+adapted from SignUpModal and SignUpForm.
The Yup validation schema for Formik is simpler.

We will also need to 'doGetWithBasicAuthCredentials' were the credentials
are just the values from the form inputs.

Differently from the SignUpForm, the response for this GET fetch will have
different status codes.
The improved handling of these response statuses (i.e. doing something
better than displaying an alert and clearing the form) remains
a TODO for the future.