Update the /api/donate/as/anon route
Now, we return a dictionary/map of string-values. Currently,
it only returns a single field: 'client_secret'

No longer we update the DB in this method, rather we create a
PaymentIntent for Stripe (and return the client secret so that
the client may finalize it)

We do include as metadata the projectId to which this payment
corresponds (i.e. which project is being donated to)
Later, this will also need to include the corresponding Stripe
Connect ID for said project's payment account.
(i.e. for now, all payments are made to a single account, the
sandbox one; a full-fledged crowdfunding platform will need to
connect the donor's account to the project's account)