Add an event Listener
It's actually a webhook. Based on the types of events POSTed to
'/api/stripe/listen' we can add side-effects for the server logic.

We are mainly concerned with updating the Donation objects/table
upon payment success.
(SEE the handlePaymentSuccess method! Note: we get the projectId
from the metadata set beforehand!!)

Using a listener is recommended in favor of a callback from the
client as the user may close the app before the callback is
resolved. Instead, the Stripe api will send the events from a
persisted log.

Stripe uses GSON to deserialize the JSON and get the event type.
It is necessary to set up the forwarding of events to this
endpoint! (in the Stripe dashboard when the server is deployed,
or via the Stripe CLI on 'localhost')