Add an AmazonS3FileTransferService
It just has the basic upload/download methods.
Upload needs an InputStream.
The last two args of the function can be given as
'upload(... ... null, null)' in the naive case.
The map may hold metadata to tag files in the bucket if
we need to preserve some domain-info associated with said files.

(There should be a better way to pass complex structs -- with no
data -- than simply passing null, but an additional
object to wrap these last 2 arguments and a builder for that
object seemed too much.)

Content-Length should be obtained from the request HTTP headers.

The download() method gets the whole file in local memory as an
array of bytes.

The corresponding bucketname has been added to application.yaml