Add an ImageService implementing the Facade
It needs the AmazonS3FileTransferService and the ImageDataRepo to
implement the functions.

Here we see why we had the (folder, filename) structure.
We can cluster files for same projectName in the same folder.
(eventually we can use some other field available in the multipart
FormData request)

There is no metadata yet, but we put the Content-Type header which
should be treated separately (custom metadata is prefixed
differently in the headers) in the awsFileTransfer service.

Download is quite simple.