Refactor ImageService
Now we make explicit that when we save a file it must be a proper image file
This means making use of its MimeType, as before.

The updateMetadataWithContentTypeAndReturnIt() method was introduced.

We throw a TempStorage exception to specify that the MultiPartFile
access to an InputStream has failed.
We would throw an ImproperFileException in the future if the file is not
of the correct image type.