(incomplete) Add a very basic AddProjectForm
It doesn't use Formik. It relies on a classic <form/>

It would create a new FormData() object when 'onSubmit' is reached.
This object would be populated with the form's input.name:value
pairs.
It would then make use of the MultiPart POST fetch to create a new
Project server-side.

Since we are using the event.target.value we don't use React
hooks for the form itself.

NOTE: the submit button would enter a loading state and the form
would only close once this loading state is over.
(The fetch promise resolves after a significant wait time.)

This is better done with ReactQuery or the equivalent
'useMutation' from Redux Modern Toolkit